mudra/requirements.txt
PyQt5>=5.15
opencv-python>=4.8
mediapipe>=0.10
numpy>=1.24
torch>=2.0
scikit-learn>=1.3
matplotlib>=3.7
pandas>=2.0
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.7
PyYAML>=6.0
